splitting FlexBox, changed it's name to ButtonGrid and EditableButton
